MI5-65 AccountListItem je prebačena u package list_components
Po review od Nino Lončara na PR.
