+++ app/src/main/java/foi/air/szokpt/ui/components/list_components/AccountListItem.kt
@@ -1,0 +1,50 @@
+package foi.air.szokpt.ui.components.list_components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import foi.air.szokpt.models.ListedAccountInformation
+import foi.air.szokpt.ui.theme.BGLevelThree
+import foi.air.szokpt.ui.theme.TextGray
+import foi.air.szokpt.ui.theme.TextWhite
+
+@Composable
+fun AccountListItem(account: ListedAccountInformation) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 4.dp, horizontal = 6.dp)
+            .background(
+                color = BGLevelThree,
+                shape = RoundedCornerShape(10.dp)
+            )
+            .clickable { println("Selected account: $account") } // Forward HERE
+            .padding(6.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Column {
+            Text(
+                text = account.role.name + " @" + account.userName,
+                color = TextGray,
+                fontSize = 12.sp,
+                fontWeight = FontWeight.SemiBold
+            )
+            Text(
+                text = "${account.name} ${account.lastName}",
+                color = TextWhite,
+                fontSize = 16.sp
+            )
+        }
+    }
+}
